backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0
greenlet==3.0.3

# ── Auth & Security ───────────────────────────────────────────────────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.7

# ── Settings & Env ────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.27.0
aiofiles==23.2.1

# ── Rate Limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── Caching ───────────────────────────────────────────────────────────────────
redis==5.0.4

# ── Market Data ───────────────────────────────────────────────────────────────
yfinance==0.2.38
pandas==2.2.2

# ── Background Tasks ──────────────────────────────────────────────────────────
apscheduler==3.10.4

# ── Backtesting / Analytics ───────────────────────────────────────────────────
polars==0.20.23
numpy==1.26.4

# ── WebSocket ─────────────────────────────────────────────────────────────────
websockets==12.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
